Begin integrating nueral network into ai

+++ src/Matrix.java
@@ -3,18 +3,18 @@ import java.util.Random;
 public class Matrix {
     public static Random random = new Random(0);
 
-    private double[][] data;
+    private float[][] data;
 
     public Matrix(int rows, int cols){
-        this.data = new double[rows][cols];
+        this.data = new float[rows][cols];
     }
 
-    public Matrix(double[][] data)
+    public Matrix(float[][] data)
     {
         this.data = data;
     }
 
-    public void setData(double[][] data)
+    public void setData(float[][] data)
     {
         this.data = data.clone();
         //TODO test pass by value
@@ -25,7 +25,7 @@ public class Matrix {
         int rows = a.data.length;
         int cols = a.data[0].length;
 
-        double[][] sum = new double[rows][cols];
+        float[][] sum = new float[rows][cols];
 
         for (int r = 0; r < rows; r++) {
             for (int c = 0; c < cols; c++) {
@@ -41,7 +41,7 @@ public class Matrix {
         int rows = a.data.length;
         int cols = a.data[0].length;
 
-        double[][] sum = new double[rows][cols];
+        float[][] sum = new float[rows][cols];
 
         for (int r = 0; r < rows; r++) {
             for (int c = 0; c < cols; c++) {
@@ -58,7 +58,7 @@ public class Matrix {
         int cols2 = b.data[0].length;
         int rows2 = b.data.length;
 
-        double[][] ret = new double[rows1][cols2];
+        float[][] ret = new float[rows1][cols2];
 
         for (int r = 0; r < rows1; r++) {
             for (int c = 0; c < cols2; c++) {
@@ -71,13 +71,13 @@ public class Matrix {
         return new Matrix(ret);
     }
 
-    public static Matrix randomMatrix(int rows, int cols, double mean, double std)
+    public static Matrix randomMatrix(int rows, int cols, float mean, float std)
     {
-        double[][] ret = new double[rows][cols];
+        float[][] ret = new float[rows][cols];
 
         for (int r = 0; r < rows; r++) {
             for (int c = 0; c < cols; c++) {
-                ret[r][c] = random.nextGaussian();
+                ret[r][c] = (float) random.nextGaussian();
             }
         }
 
@@ -109,6 +109,6 @@ public class Matrix {
         return ret;
     }
 
-    public double[][] getData() { return this.data; }
+    public float[][] getData() { return this.data; }
 
 }
